+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Removed
 
 - Remove `setAppElement` from static on `Modal` and `ModalDialog` in favor of the `appElement` prop on `Confirm`, `ConfirmDelete`, `Modal`, and `ModalDialog`. Defaults to `document.body`.
+- The ability to specify the `tag` prop for the `DoorEffect` component.
 
 ## [15.0.1] - 2019-08-28